Create Todo.java
I had an error where I tried to move my Todo.java from the default package to the project-level, but I had already committed the Todo.java file before. It kept saying that the remote contained work that I didn't have locally, and when I tried to move it back it said the same thing. I still have Todo.java in my eclipse workspace, so I've copied and pasted it into this file through simply creating a file on GitHub.